[Jon Willeke] I'm attaching a few tweaks to build psqlodbc with Unicode support on Unix.  I changed some code that assumes two-byte SQLWCHAR and I changed configure.ac and Makefile.am, as necessary.
I tested this with unixODBC 2.2.7 on SuSE Linux 8.2.  I built unixODBC --with-iconv-ucode-enc=UCS-4LE and with SQL_WCHART_CONVERT defined, so that SQLWCHAR is wchar_t.  I have not tested it with UCS-2 / unsigned short.

I also tested this with iODBC 3.51.1, which always uses wchar_t on Unix.

One caveat, if anyone actually plays with this: sqlext.h should include sqlucode.h, so I haven't patched psqlodbc to do so.  The next release of unixODBC will contain this fix, and I've mentioned it to the iODBC maintainer.