Fix bug with VARCHAR(5) parameter and UseServerSidePrepare=0
If you passed a VARCHAR parameter with column_size 5 to SQLPrepare, and you
had UseServerSideprepare=0 and BoolsAsChar=1, the code describes the
parameters before running the query, so that it can check if the parameter
is a boolean. That produced a bogus extra empty result set to the caller in
subsequent SQLExecute().

This fixes the two regression test failures.

Patch by Hiroshi Inoue, with minor copy-editing by me.