Change the default for UseServerSidePrepare to 1.
The docs have recommended UseServerSidePrepare=1 for server versions 7.4
onwards, so it seems prudent to change the default so that people don't need
to remember to specify it manually.

With UseServerSidePrepare=1, the "insertreturning" regression test case
behaves better, ie. SQLNumResultCols() correctly returns the number of
columns for an INSERT RETURNING statement, even when called before
SQLExecute(). Hence change the expected output. I'm not 100% if the behavior
with UseServerSidePrepare=0 was even correct, so I'm not adding an
alternative expected output file for it.