Rip out broken retry/timeout logic in SOCK_wait_for_ready.
At a quick glance, the logic looked like retry with progressive timeout, but
it was horribly broken. First of all, the timeout was only used if
retry_count was passed as 0. And in that case, the timeout was 0. So
retry_count == 0 actually meant "check if data can be read/written without
blocking", while retry_count > 0 meant "sleep until data is available". There
was also some code to handle retry_count < 0, but that was dead code because
none of the callers passed a negative value.

The hardcoded max. retry count of 30 retries was quite bogus too. Filling
the output buffer more 30 times is not necessary an error, and it's not
clear that the ssl library could not return SSL_ERROR_WANT_READ or
SSL_ERROR_WANT_WRITE more than 30 times. Unlikely, yes, but possible.

So, just rip out the retry_count and progressive timeout logic. Replace
the retry_count argument with a simple "should block?" argument. I didn't
dare to remove the condition that we always wait if SSL is enabled, although
I don't understand that at all. But perhaps there's some well-hidden reason
for that.

Per report from Prakash Itnal, although this just removes the broken code,
and won't enable the kind of a timeout that he wanted.