Fix crash if connection is closed while sending a query to the server.
This is a similar bug that was fixed in commit 22b151eb. Some functions
reset conn->sock to NULL on a socket error, so we must not keep a copy of
it in a local variable. Also, SOCK_get_id should check for NULL argument
like all the rest of the SOCK_get/put_* functions.