Minor code cleanup.
The main change is to move the responsibility of adding the "length" header
to the QueryBuild from ResolveOneParam to its caller, when constructing the
parameters to send over to the server in a Bind message. ResolveOneParam
is complicated enough without that, so it seems more clear this way.

The rest is just fixes to the macros to make them deal with a "&qb" argument,
by adding parens around the macro argument. And while at it, clean up the
whitespace in them.