Additonal improvement on MSDTC enlistment.
Unfortunately the current pgxalib.dll can't recover transactions which use
sslmode verify-[ca|full] or whose authentication is SSPI, certificate or
ident.
When MSDTC.exe tries to connect to PostgreSQL it does so under Windows user
NETWORKSERVICE, which won't match the username the original user connected
to PostgreSQL as. So PostgreSQL will reject the connection.

In addition XARMcreate() doesn't detect it on enlistment because invoked
xa_open() doesn't connect to the specified datasource. Though it has no
problem unless crash occurs between PREAPARE and COMMIT, the recovery of
the transaction is impossible when such a crash occurs.

Now 3 diffrent behaviors can be speicified using the xaopt connection
string option.

 1: The current default behavior.
 2: Same as above except it simply examines if the recovery for the specified
    datasource is unavailable before calling XARMcreate(). Currently it only
    examines whether the connection uses sslmode verify-[ca|full]. I can't find
    the way to detect SSPI, certicate or ident authentication.
 3: Invoked xa_open() tries to connect to the specified datasource immediately
    as user Network Service.