Improve regression test for reading a bytea column to SQL_C_CHAR.
When converting a bytea to SQL_C_CHAR, the driver uses upper-case letters in
the hex output, while the backend uses lower-case letters. That leads to
different output when using bytea_output=hex than with bytea_output=escape.
Set bytea_output explicitly in the regression test, so that it works
regardless of the current bytea_output setting. And test both settings - the
code path taken in the driver is quite different.