Fix max identifier length stuff.
CC_get_max_idlen function didn't work correctly. It sent the "show
identifier_length" query to the backend, but assumed that the value comes
back in the command tag. It doesnt; the command tag is always "SHOW".

We fell back to the compiled in NAMEDATALEN_V73 default. But that's
off-by-one; the actual max identifier length is NAMEDATALEN - 1.