Run the odbc[inst].ini generation script in a more Makefile fashion.
The expected output didn't match what I got on my laptop. Not sure why,
or if it worked for anyone, but in any case it's more straightforward to
let make run the script directly, rather than through pg_regress.

In the passing, allow passing extra command line options to odbcini-gen.sh,
which will be appended to the generated odbc.ini. This makes it handy to
run the regressions with extra non-default options.