Add test case for SQLSetCursorName and SQLGetCursorName.
The test case also contains tests for UPDATE ... WHERE CURRENT OF, but
it's disabled because the driver doesn't actually support it. I wrote the
test before I realized that, and didn't want to throw it away, in case we
ever want to add support for it.

It also contains an #ifdef'd out test for cursor names that contain quotes.
The driver doesn't handle that correctly, but I didn't want to fix it right
now. No sane application does that anyway.