Silence regression failure of diagnostic test.
Now that I fixed the bug in the error handler, we're not losing the libpq's
message about lost connect anymore. It overwrites the previous FATAL error,
which is unfortunate, but it's been long-standing behaviour that if multiple
errors are returned as result for a query, only the last one is kept.