Refactor and add comments for the modes ResolveOneParam can operate in.
ResolveOneParam can work in four different modes that are mutually
exclusive. Put the mode in an enum field in the QueryBuild struct,
separately from the additional flags that can be set. Add comments to
explain what each mode does.

Refactor the way CVT_SPECIAL_CHARS works with the query buffer along the
way. It was always a bit awkward, but with this refactoring it would've
become more so, as the mode would've needed to be another argument to it.