Fix bug in detection of E'' literals.
The code assigned to escape_in_literal local variable when it saw a quote
characeter after E. And checked the local variable on subsequent iterations.
But the function is called separately for each processed characters, so the
local variable was reset between each call. Move the flag to the QueryParse
struct so that it survives across calls.

Expand the 'quotes' test case to cover this.

Clang Static Analyzer warned that the escape_in_literal variable was set
but not used, which lead me into looking at this more closely.