Let the server decide if SQL_LONGVARBINARY means a bytea or large object.
This reverts the behaviour of SQL_LONGVARBINARY parameters to the way it was
before commit 1ccd04c70b8037b268eefcaae09ce131c8c33b59. We can't reliably
guess in the driver whether the parameters should be sent to the server
as 'bytea' or as a large object, so send it as 'unknown' and let the server
decide.

Diagnosis and original patch by Hiroshi Inoue, I only added the comment.