Fix handling of return value in {? = CALL ... } syntax.
When sending the query to the server, need to use the number of parameters
calculated by build_libpq_bind_params(), which excludes the OUT parameter,
rather than the original number of parameters. The branch that executes
an already prepared statement got this right.

There was one test query for the CALL syntax already, but it was commented
out, because it doesn't work for unrelated issues. Add a simpler test case
that does work.

This was broken by commit e6d7ded, in version 09.03.0400. The code was
heavily refactored in commit e85fbb24, but that refactoring failed to fix
this issue.

Reported by Haribabu Kommi.