Use SIZEOF_LONG_INT to check whether SQLROWSETSIZE exists.
I changed this originally to use SIZEOF_LONG_INT, in commit 661b9bf8. That
commit also changed an unrelated check in psqlodbc.h, but that was wrong.
Hiroshi fixed that in commit 19b3dee3, by changing them to use SIZEOF_LONG
instead. That was correct for psqlodbc.h, but was in turn wrong for
odbcapi.c.

SIZEOF_LONG_INT is defined in UnixODBC headers, while SIZEOF_LONG is defined
by psqlodbc's ./configure. In odbcapi.c, we use it to determine what format
modifier, %ld or %lld, to use for pointers. (Why don't we use %p?). We should
rely on our autoconf check for that. In odbcapi.c, however, the check should
match that in the ODBC headers, to determine whether SQLROWSIZE exists, so
we should use SIZEOF_LONG_INT.