Increase WiX-5-ness
- Remove most Component elements in favor of naked files
- Replace PGFOLDER variable with ProgramFiles6432Folder

The Condition element on Component has become an attribute in WiX 5, and
the combination of preprocessor and component condition (i.e. include
a non-x64 condition only in a non-x64 package) to install PGXA only from
a package native to the underlying OS does not work anymore. Replace it
with a pure component condition that includes the package architecture.

The Template summary property has this information already, but it is
not available from within the installation, and I'm not about to start
writing custom actions.