Use WiX 5 (#66)
* Make product code automatically random

Instead of externally generating it and using it explicitly.

* Componentize the merge module

This prevents upgrades from leaving files from the previous version
behind that are not in the new version.

* Configure drivers on x64 The Way It Should Be Done

Use the ODBCDriver element only, without additional Registry settings.
This works, although it does not solve the file name issue and therefore
still requires removing the short file names from the File table.

It also works for installing the x86 driver on an x64 system. Whether it
works on an x86 system remains unknown.

* Configure drivers on x86 In The Exact Same Way

Tested on Windows 10 x86; the driver installs correctly and works
at first glance (in Access, to be precise). Registry and file system
look good.

The comment saying it "probably would be safe" is probably correct.

* Use candle's -arch option to select build platform

According to docs, using Package/@Platform "is discouraged". The -arch
option also provides the default bitness for components, making the
Component/@Win64 attribute unnecessary in single-architecture packages.

* Reindent

* Save some attributes

* Replace upgrade logic with MajorUpgrade element

This element compiles into approximately the same entries in the
Upgrade, CustomAction, and InstallExecuteSequence tables.

* Remove unused GUID

I think this was the product code, a long time ago.

* Make MSI file smaller

By about 20 percent.

* Better select where to remove short file names

Use the FileName column itself instead of the file ID. This is less
susceptible to changes in file IDs (and component IDs) that may happen
when relying more on default behaviors in future WiX versions.

* Build with WiX 5

This is a preview. It appears to build a working x64 installer; I have
tested nothing else yet.

* Clean up modify_msi.vbs

Also, put blame where it belongs, which is Windows Installer. If this
was a bug in WiX, then the fix would be to enforce 8.3 file names for
ODBC drivers, and we wouldn't like that very much either, would we?

* Increase WiX-5-ness

- Remove most Component elements in favor of naked files
- Replace PGFOLDER variable with ProgramFiles6432Folder

The Condition element on Component has become an attribute in WiX 5, and
the combination of preprocessor and component condition (i.e. include
a non-x64 condition only in a non-x64 package) to install PGXA only from
a package native to the underlying OS does not work anymore. Replace it
with a pure component condition that includes the package architecture.

The Template summary property has this information already, but it is
not available from within the installation, and I'm not about to start
writing custom actions.

* Update build workflow for WiX 5